LayoutTests/fast/animation/request-animation-frame-missing-arguments-expected.txt

Test how animation methods react to too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS webkitRequestAnimationFrame() threw exception TypeError: Not enough arguments.
PASS webkitCancelRequestAnimationFrame() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
